feat: skip even looking up information on items we know are done
This actually only happens once you've run across the item *twice*: we
don't have a good synchronization process to know that you've finished
all files of an item while handling each file one at a time, but we can
do it when we're looking at all of the item's files.

Basically, this speeds things up if you're starting and stopping
repeatedly on the same collection, but not if you just do it once.

Also, you will still have to page through the search results to catch
back up to where you left off, which may take some time.